fix(debugger/ios): do not prompt for certificate info on attach request
