+++ src/debugger/titaniumSourceMapTransformer.ts
@@ -1,8 +1,8 @@
 import * as fs from 'fs-extra';
 import * as path from 'path';
 import { BaseSourceMapTransformer } from 'vscode-chrome-debug-core';
+import { TitaniumAttachRequestArgs, TitaniumLaunchRequestArgs } from '../common/extensionProtocol';
 import { determineProjectType } from '../common/utils';
-import { TitaniumAttachRequestArgs, TitaniumLaunchRequestArgs } from './titaniumDebugAdapter';
 
 export class TitaniumSourceMapTransformer extends BaseSourceMapTransformer {
